% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bredutils.R
\name{bread}
\alias{bread}
\title{Helper to read in bed file
Contributes not much except setting first 3 column names}
\usage{
bread(bedfile, ...)
}
\arguments{
\item{bedfile}{A path to a bed file}

\item{...}{Additional arguments passed to data.table::fread()}
}
\value{
A data table object
}
\description{
\code{bread} reads in a bed file and sets the first three column names.
}
\details{
The returned object is in data.table format because it's better that way
}
