% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createA.R
\name{createA}
\alias{createA}
\title{Given the vectors of sire and dam, return additive relationship matrix 'A'}
\usage{
createA(s, d)
}
\arguments{
\item{s}{a vector of sire}

\item{d}{a vector of dam}
}
\value{
additive relationship matrix 'A'
}
\description{
Given the vectors of sire and dam, return additive relationship matrix 'A'
}
