% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAinvL.R
\name{createAinvL}
\alias{createAinvL}
\title{Given the vectors of sire and dam, directly return inverse of additive relationship matrix 'A' based on a lower traingular matrix 'L' without creating the 'A' itself.}
\usage{
createAinvL(s, d)
}
\arguments{
\item{s}{a vector of sire}

\item{d}{a vector of dam}
}
\value{
inverse of additive relationship matrix 'A'
}
\description{
Given the vectors of sire and dam, directly return inverse of additive relationship matrix 'A' based on a lower traingular matrix 'L' without creating the 'A' itself.
}
