% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createD.R
\name{createD}
\alias{createD}
\title{Given the vectors of sire and dam, return a dominance relationship matrix 'D'}
\usage{
createD(s, d)
}
\arguments{
\item{s}{vector of sire}

\item{d}{a vector of dam}
}
\value{
a dominance relationship matrix 'D'
}
\description{
Given the vectors of sire and dam, return a dominance relationship matrix 'D'
}
