% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createL.R
\name{createL}
\alias{createL}
\title{Given the vectors of sire and dam, return a lower triangular matrix 'L' such that LL' = 'A'}
\usage{
createL(s, d)
}
\arguments{
\item{s}{a vector of sire}

\item{d}{a vector of dam}
}
\value{
a lower triangular matrix 'L'
}
\description{
Given the vectors of sire and dam, return a lower triangular matrix 'L' such that LL' = 'A'
}
