% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{custom_array_sort}
\alias{custom_array_sort}
\title{stage sort by custom_peak_sort function}
\usage{
custom_array_sort(x, seq_array)
}
\arguments{
\item{x}{an array of numbers}

\item{seq_array}{an array of numbers}
}
\value{
an sorted array of numbers by custom_peak_sort function
}
\description{
stage sort by custom_peak_sort function
}
\examples{
custom_array_sort(c(1:10),c(5,5))
}
