% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{custom_peak_sort}
\alias{custom_peak_sort}
\title{peak sort}
\usage{
custom_peak_sort(x)
}
\arguments{
\item{x}{an array of numbers}
}
\value{
an array of numbers
}
\description{
peak sort
}
\examples{
custom_peak_sort(1:5)
custom_peak_sort(20:30)
}
