% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{custom_rnorm}
\alias{custom_rnorm}
\title{custom gauss sampling}
\usage{
custom_rnorm(n, mean, sd, seed = 2024)
}
\arguments{
\item{n}{a number of sample}

\item{mean}{mean of sample}

\item{sd}{sd of sample}

\item{seed}{random of seed}
}
\value{
an array of sample
}
\description{
custom gauss sampling
}
\examples{
custom_rnorm(100,10,1)
}
