% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{custom_sample}
\alias{custom_sample}
\title{custom average sampling}
\usage{
custom_sample(x = x, size = size, replace = FALSE, seed = 2024)
}
\arguments{
\item{x}{an array of sample}

\item{size}{number of sample}

\item{replace}{boolean value of repeat}

\item{seed}{random of seed}
}
\value{
an array of sample
}
\description{
custom average sampling
}
\examples{
custom_sample(c(1:10),10,F)
}
