% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emreml.R
\name{emreml}
\alias{emreml}
\title{Given the components of MME, estimate variance components by EM Algorithm.}
\usage{
emreml(Ainv, y, X, Z, initE, initU, disp)
}
\arguments{
\item{Ainv}{an inverse of additive relationship matrix}

\item{y}{a vector of response}

\item{X}{an incidence matrix for fixed effects}

\item{Z}{an incidence matrix for random effects}

\item{initE}{initial value for the residual variance}

\item{initU}{initial value for the additive genetic variance}

\item{disp}{a logical value. If true, display estimates of variance components at each iteration.}
}
\value{
a vector of the residual variance and the additive genetic variance
}
\description{
Given the components of MME, estimate variance components by EM Algorithm.
}
