% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jor.R
\name{jor}
\alias{jor}
\title{Given the MME, iteratively solve for the solutions by Jocobi Over Relaxation.}
\usage{
jor(LHS, RHS, inits, w, disp)
}
\arguments{
\item{LHS}{left hand side of MME}

\item{RHS}{right hand side of MME}

\item{inits}{a vecotr of initial values for the solutions}

\item{w}{a vector of relaxation factor}

\item{disp}{a logical value. If true, show the solutions at each iteration}
}
\value{
a vector of iteratively solve for the solutions
}
\description{
Given the MME, iteratively solve for the solutions by Jocobi Over Relaxation.
}
