% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{legendre}
\alias{legendre}
\title{coefficient matrix (lambda) of n-th order Legendre polynomials}
\usage{
legendre(n, gengler)
}
\arguments{
\item{n}{order of polynomials}

\item{gengler}{logical value}
}
\value{
coefficient matrix (lambda)
}
\description{
coefficient matrix (lambda) of n-th order Legendre polynomials
}
