% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{pen_simu}
\alias{pen_simu}
\title{batch simulation of piggery-unit-pen encoding}
\usage{
pen_simu(
  piggery_id,
  piggery_begin,
  piggery_end,
  unit_begin,
  unit_end,
  pen_begin,
  pen_end,
  sep = "-"
)
}
\arguments{
\item{piggery_id}{abbreviation of piggery code}

\item{piggery_begin}{begin number of piggery}

\item{piggery_end}{end number of piggery}

\item{unit_begin}{begin number of unit}

\item{unit_end}{end number of unit}

\item{pen_begin}{begin number of pen}

\item{pen_end}{end number of pen}

\item{sep}{separator}
}
\value{
an array of piggery-unit-pen encoding
}
\description{
batch simulation of piggery-unit-pen encoding
}
\examples{
pen_simu("PH",1,6,1,2,1,45)
pen_simu("PH",7,8,3,5,1,20)
}
