% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{pen_simu2}
\alias{pen_simu2}
\title{batch simulation of piggery-pen encoding}
\usage{
pen_simu2(
  piggery_id,
  piggery_begin,
  piggery_end,
  pen_begin,
  pen_end,
  sep = "-"
)
}
\arguments{
\item{piggery_id}{abbreviation of piggery code}

\item{piggery_begin}{begin number of piggery}

\item{piggery_end}{end number of piggery}

\item{pen_begin}{begin number of pen}

\item{pen_end}{end number of pen}

\item{sep}{separator}
}
\value{
an array of piggery-pen encoding
}
\description{
batch simulation of piggery-pen encoding
}
\examples{
pen_simu2("PH",1,4,1,60)
}
