% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{piggery_simu}
\alias{piggery_simu}
\title{batch simulation dataframe of piggery information}
\usage{
piggery_simu(
  piggery_type_id,
  piggery_begin,
  piggery_end,
  piggery_name_begin,
  piggery_name_end,
  piggery_type,
  branch_id,
  sep = "-"
)
}
\arguments{
\item{sep}{}
}
\value{
a dataframe of piggery information
}
\description{
batch simulation dataframe of piggery information
}
\examples{
piggery_simu("PH",2,18,"配怀","栋","tmp","pigfarm")
}
