% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quass.R
\name{quaas}
\alias{quaas}
\title{Given the vectors of sire and dam, directly return inverse of additive relationship matrix 'A' without creating the 'A' itself}
\usage{
quaas(s, d)
}
\arguments{
\item{s}{a vector of sire}

\item{d}{a vector of dam}
}
\value{
inverse of additive relationship matrix 'A'
}
\description{
Given the vectors of sire and dam, directly return inverse of additive relationship matrix 'A' without creating the 'A' itself
}
