% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdtime.R
\name{stdtime}
\alias{stdtime}
\title{Given time points covariate and order of fit for Legendre polynomials, return matrix 'M' containing the polynomials of standardized time. 'M' is order t (number of time points) by k (order of Legendre polynomials).}
\usage{
stdtime(t, n, tmax, tmin)
}
\arguments{
\item{t}{a vector of time, age or days}

\item{n}{order of polynomials}

\item{tmax}{max time (optional)}

\item{tmin}{min time (optional)}
}
\value{
matrix 'M' containing the polynomials of standardized time
}
\description{
Given time points covariate and order of fit for Legendre polynomials, return matrix 'M' containing the polynomials of standardized time. 'M' is order t (number of time points) by k (order of Legendre polynomials).
}
