% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{piggery_simu}
\alias{piggery_simu}
\title{batch simulation dataframe of piggery information}
\usage{
piggery_simu(
  piggery_type_id,
  piggery_begin,
  piggery_end,
  piggery_name_begin,
  piggery_name_end,
  piggery_type,
  branch_id,
  sep = "-"
)
}
\arguments{
\item{piggery_type_id}{abbreviation of piggery code}

\item{piggery_begin}{begin number of piggery}

\item{piggery_end}{end number of piggery}

\item{piggery_name_begin}{begin name of piggery}

\item{piggery_name_end}{end name of piggery}

\item{piggery_type}{abbreviation of piggery type code}

\item{branch_id}{digital code of branch}

\item{sep}{separator}
}
\value{
a dataframe of piggery information
}
\description{
batch simulation dataframe of piggery information
}
\examples{
piggery_simu("PH",2,18,"配怀","栋","tmp","pigfarm")
}
